fixing gcc complaints in tests

+++ ambient/regression/functional/aux.hermitian.cpp
@@ -8,10 +8,8 @@ BOOST_AUTO_TEST_CASE_TEMPLATE( CONJ_INPLACE, T, test_types)
     pMatrix pA(T::valuex, T::valuey);
     sMatrix sA(T::valuex, T::valuey);
 
-    if(T::valuex == T::valuey)
-        generate_hermitian(pA);
-    else
-        generate(pA);
+    if((int)T::valuex == (int)T::valuey) generate_hermitian(pA);
+    else generate(pA);
 
     sA = cast<sMatrix>(pA);
